Don't run Lighthouse CI on Depfu PRs
Depfu is a bot that automatically makes PRs for updating dependencies. It does not trigger a staging build (at least not right now, that might be worth trying in the future though). But since it never triggers a staging build that ALSO means Lighthouse CI hangs for several hours waiting for the site to be deployed, until it finally times out. (You'll see this error in the checks section of all the previous Depfu PRs)

This PR fixes that by disabling the Lighthouse run for depfu tagged PRs.